Optimized the main loop
Shared memory values are now FP32, this will reduce precision slightly
but as long as we accumulate in a FP64 value it shouldn't be too bad.
This allows better performance on architectures which do not have as
much shared memory as the Tesla K80.

Removed some unneeded logic that was creating branching complexity

Partially unrolled inner loop to reveal some ILP that was originally
not present